+++ doc/change-log.md
@@ -31,6 +31,7 @@
 - 新增 `QryTermCmdReq` 类
 - 新增 `info.gratour.jt808common.JT808Consts.ProtocolVersions.detectProtoVerBySimNoLen` 方法
 - `OpenStrmReq` 去除 `talkSendProtoVer` 属性
+- `StrmStatus` 增加 `sent`, `sendRate` 属性
 
 ## 不兼容变更
 - `AvUploadReq` 增加 `fileSz` 必填属性